backend/requirements.txt
flask
openai
gtts
psycopg2-binary
python-dotenv

